requirements.txt
numpy==1.24.3
Pillow==10.0.0
customtkinter==5.2.0
pyinstaller==6.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2 